modify fetching data structure.
